requirements.txt
Django==5.2.1
pillow==11.2.1
requests==2.32.3
djangorestframework==3.16.0
django-filter==25.1

gunicorn==23.0.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
whitenoise==6.9.0